[CAZB-2143] Separate manage user endpoints
