+++ app/views/users_management/remove_users/delete.haml
